+++ MGEX_BackOffice_EF6/Models/ProductLast.cs
@@ -13,7 +13,7 @@ namespace MGEX_BackOffice_EF6.Models
         public decimal O_Price { get; set; }
         public int B_Seq_Num { get; set; }
         public int O_Seq_Num { get; set; }
-        public int B_Time { get; set; }
-        public int O_Time { get; set; }
+        public decimal B_Time { get; set; }
+        public decimal O_Time { get; set; }
     }
 }
